rename index to orientation
